mostly working in the curtain lights - there's one little tiny bug where "off" is actually very very dim but still "on"
